core/users/serializers.py
from rest_framework import serializers

from core.users.models import Author


class AuthorSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Author
        fields = ('id', 'first_name', 'last_name', 'email', 'github', 'bio')
